+++ src/main/java/com/stcom/smartmealtable/domain/Address/Address.java
@@ -1,4 +1,4 @@
-package com.stcom.smartmealtable.domain.common;
+package com.stcom.smartmealtable.domain.Address;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -7,11 +7,13 @@ import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import java.math.BigDecimal;
+import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
+@NoArgsConstructor
 public class Address {
 
     @Id
@@ -27,15 +29,28 @@ public class Address {
 
     private String alias;
 
-    private BigDecimal latitude;
+    private Double latitude;
 
-    private BigDecimal longitude;
+    private Double longitude;
 
     @Enumerated(EnumType.STRING)
     private AddressType type;
 
+    @Builder
+    public Address(String lotNumberAddress, String roadAddress, String detailAddress, String alias, Double latitude,
+                   Double longitude, AddressType type) {
+        this.lotNumberAddress = lotNumberAddress;
+        this.roadAddress = roadAddress;
+        this.detailAddress = detailAddress;
+        this.alias = alias;
+        this.latitude = latitude;
+        this.longitude = longitude;
+        this.type = type;
+    }
+
+
     public void updateAddress(String lotNumberAddress, String roadAddress, String detailAddress,
-                              BigDecimal latitude, BigDecimal longitude) {
+                              Double latitude, Double longitude) {
         this.lotNumberAddress = lotNumberAddress;
         this.roadAddress = roadAddress;
         this.detailAddress = detailAddress;
