+++ src/main/java/com/stcom/smartmealtable/domain/Address/AddressType.java
@@ -1,4 +1,4 @@
-package com.stcom.smartmealtable.domain.common;
+package com.stcom.smartmealtable.domain.Address;
 
 public enum AddressType {
     HOME, SCHOOL, OFFICE
