+++ src/main/java/com/stcom/smartmealtable/domain/Budget/Budget.java
@@ -1,6 +1,6 @@
 package com.stcom.smartmealtable.domain.Budget;
 
-import com.stcom.smartmealtable.common.BaseTimeEntity;
+import com.stcom.smartmealtable.domain.common.BaseTimeEntity;
 import com.stcom.smartmealtable.domain.member.Member;
 import jakarta.persistence.Column;
 import jakarta.persistence.DiscriminatorColumn;
