--- src/main/java/com/stcom/smartmealtable/domain/member/MemberService.java
@@ -1,24 +1,0 @@
-package com.stcom.smartmealtable.domain.member;
-
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class MemberService {
-
-    private final MemberRepository memberRepository;
-
-    public boolean isEmailExists(String email) {
-        List<Member> findMembers = memberRepository.findMemberByEmail(email);
-        if (findMembers.isEmpty()) {
-            return false;
-        }
-        return true;
-    }
-
-    public void saveMember(Member member) {
-        memberRepository.save(member);
-    }
-}
