+++ src/main/java/com/stcom/smartmealtable/infrastructure/social/KakaoHttpMessage.java
@@ -1,12 +1,12 @@
-package com.stcom.smartmealtable.web.auth.social;
+package com.stcom.smartmealtable.infrastructure.social;
 
-import static com.stcom.smartmealtable.web.auth.social.SocialConst.KAKAO;
+import static com.stcom.smartmealtable.infrastructure.social.SocialConst.KAKAO;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import com.stcom.smartmealtable.web.dto.token.TokenDto;
+import com.stcom.smartmealtable.service.dto.token.TokenDto;
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
