+++ src/main/java/com/stcom/smartmealtable/infrastructure/social/SocialAuthService.java
@@ -1,31 +1,29 @@
-package com.stcom.smartmealtable.web.auth.social;
+package com.stcom.smartmealtable.infrastructure.social;
 
-import static com.stcom.smartmealtable.web.auth.social.SocialConst.GOOGLE;
-import static com.stcom.smartmealtable.web.auth.social.SocialConst.KAKAO;
+import static com.stcom.smartmealtable.infrastructure.social.SocialConst.GOOGLE;
+import static com.stcom.smartmealtable.infrastructure.social.SocialConst.KAKAO;
 
-import com.stcom.smartmealtable.web.dto.token.TokenDto;
+import com.stcom.smartmealtable.service.dto.token.TokenDto;
 import jakarta.annotation.PostConstruct;
+import jakarta.validation.constraints.NotEmpty;
 import java.util.HashMap;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestClient;
-import org.springframework.web.client.RestClient.RequestBodySpec;
 import org.springframework.web.client.RestClient.ResponseSpec;
 
-@Component
+@Service
 @RequiredArgsConstructor
-public class SocialHttpMessageManager {
+public class SocialAuthService {
 
     private final Map<String, SocialHttpMessage> socialMap = new HashMap<>();
     private final KakaoHttpMessage kakaoHttpMessage;
     private final GoogleHttpMessage googleHttpMessage;
+    private final RestClient client = RestClient.create();
 
-    public RequestBodySpec getTokenRequestMessage(RestClient client, String provider, String code) {
-        return socialMap.get(provider).getRequestMessage(client, code);
-    }
-
-    public TokenDto getTokenResponse(ResponseSpec responseSpec, String provider) {
+    public TokenDto getTokenResponse(@NotEmpty String provider, @NotEmpty String code) {
+        ResponseSpec responseSpec = socialMap.get(provider).getRequestMessage(client, code).retrieve();
         return socialMap.get(provider).getTokenResponse(responseSpec);
     }
 
@@ -34,4 +32,6 @@ public class SocialHttpMessageManager {
         socialMap.put(KAKAO, kakaoHttpMessage);
         socialMap.put(GOOGLE, googleHttpMessage);
     }
+
+
 }
